// src/enums/keyboard.enums.ts
export enum keyboard {
  AboutUs = '🏢About us',
  Vacancies = '👨‍💻Vacancies',
  ContactUs = '☎️Contact us',
  Comment = '💬Comment',
}

export enum keyboards {
  Back = '🔙Back',
  Home = '🏠Home',
}

export enum vacboard {
  vac1 = 'Auto Transportation Broker',
  vac2 = 'vac name',
}

export enum activity {
  student = 'I am a student',
  work = 'I work',
  job = 'I am looking a new job',
}

export enum level {
  beg = 'Beginner',
  int = 'Intermedient',
  adv = 'Advanced',
  Flu = 'Fluent',
}